+++ server/src/main/java/org/jetbrains/bsp/bazel/server/bep/BepServer.java
@@ -5,6 +5,7 @@ import ch.epfl.scala.bsp4j.StatusCode;
 import ch.epfl.scala.bsp4j.TaskFinishParams;
 import ch.epfl.scala.bsp4j.TaskId;
 import ch.epfl.scala.bsp4j.TaskStartParams;
+import ch.epfl.scala.bsp4j.TextDocumentIdentifier;
 import com.google.devtools.build.lib.buildeventstream.BuildEventStreamProtos;
 import com.google.devtools.build.v1.BuildEvent;
 import com.google.devtools.build.v1.PublishBuildEventGrpc;
@@ -24,6 +25,7 @@ import java.util.AbstractMap;
 import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.Map;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.apache.logging.log4j.LogManager;
@@ -38,6 +40,8 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
 
   private final BuildClient bspClient;
 
+  private BepLogger bepLogger;
+
   private final Deque<Map.Entry<TaskId, String>> startedEvents = new ArrayDeque<>();
   private final DiagnosticsService diagnosticsService;
   private BepOutputBuilder bepOutputBuilder = new BepOutputBuilder();
@@ -47,8 +51,11 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
     this.diagnosticsService = diagnosticsService;
   }
 
-  public static BepServer newBepServer(BuildClient client, Path workspaceRoot) {
-    return new BepServer(client, new DiagnosticsService(workspaceRoot));
+  public static BepServer newBepServer(
+      BuildClient client,
+      Path workspaceRoot,
+      Map<String, Set<TextDocumentIdentifier>> hasAnyProblems) {
+    return new BepServer(client, new DiagnosticsService(workspaceRoot, hasAnyProblems));
   }
 
   public static NettyServerBuilder nettyServerBuilder() {
@@ -73,9 +80,11 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
       BuildEventStreamProtos.BuildEvent event =
           BuildEventStreamProtos.BuildEvent.parseFrom(buildEvent.getBazelEvent().getValue());
 
-      LOGGER.debug("Got event {}", event);
+      LOGGER.trace("Got event {}", event);
 
       processBuildStartedEvent(event);
+      processProgressEvent(event);
+      processBuildMetrics(event);
       processFinishedEvent(event);
       processActionCompletedEvent(event);
       fetchNamedSet(event);
@@ -100,10 +109,23 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
     }
   }
 
+  private void processProgressEvent(BuildEventStreamProtos.BuildEvent event) {
+    if (event.hasProgress()) {
+      bepLogger.onProgress(event.getProgress());
+    }
+  }
+
+  private void processBuildMetrics(BuildEventStreamProtos.BuildEvent event) {
+    if (event.hasBuildMetrics()) {
+      bepLogger.onBuildMetrics(event.getBuildMetrics());
+    }
+  }
+
   private void consumeBuildStartedEvent(BuildEventStreamProtos.BuildStarted buildStarted) {
     bepOutputBuilder = new BepOutputBuilder();
     TaskId taskId = new TaskId(buildStarted.getUuid());
     String originId = extractOriginIdFromOptions(buildStarted.getOptionsDescription());
+    bepLogger = new BepLogger(bspClient, originId);
     TaskStartParams startParams = new TaskStartParams(taskId);
     startParams.setEventTime(buildStarted.getStartTimeMillis());
 
@@ -131,7 +153,6 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
     StatusCode exitCode = ExitCodeMapper.mapExitCode(buildFinished.getExitCode().getCode());
     TaskFinishParams finishParams = new TaskFinishParams(startedEvents.pop().getKey(), exitCode);
     finishParams.setEventTime(buildFinished.getFinishTimeMillis());
-
     bspClient.onBuildTaskFinish(finishParams);
   }
 
@@ -193,8 +214,16 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
     var label = event.getId().getTargetCompleted().getLabel();
     var targetComplete = event.getCompleted();
     var outputGroups = targetComplete.getOutputGroupList();
-    LOGGER.debug("Consuming target completed event " + targetComplete);
+    LOGGER.trace("Consuming target completed event " + targetComplete);
     bepOutputBuilder.storeTargetOutputGroups(label, outputGroups);
+    if (targetComplete.getSuccess()) {
+      // clear former diagnostics by publishing an empty array of diagnostics
+      // why we do this on `target_completed` instead of `action_completed`?
+      // because `action_completed` won't be published on build success for a target.
+      // https://github.com/bazelbuild/bazel/blob/d43737f95d28789bb2d9ef2d7f62320e9a840ab0/src/main/java/com/google/devtools/build/lib/buildeventstream/proto/build_event_stream.proto#L157-L160
+      var events = diagnosticsService.clearFormerDiagnostics(label);
+      events.forEach(bspClient::onBuildPublishDiagnostics);
+    }
   }
 
   private void processAbortedEvent(BuildEventStreamProtos.BuildEvent event) {