+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/ExecuteService.kt
@@ -9,6 +9,7 @@ import ch.epfl.scala.bsp4j.RunParams
 import ch.epfl.scala.bsp4j.RunResult
 import ch.epfl.scala.bsp4j.TestParams
 import ch.epfl.scala.bsp4j.TestResult
+import ch.epfl.scala.bsp4j.TextDocumentIdentifier
 import io.grpc.Server
 import org.eclipse.lsp4j.jsonrpc.CancelChecker
 import org.eclipse.lsp4j.jsonrpc.ResponseErrorException
@@ -31,10 +32,11 @@ class ExecuteService(
     private val projectProvider: ProjectProvider,
     private val bazelRunner: BazelRunner,
     private val workspaceContextProvider: WorkspaceContextProvider,
-    private val bspClientTestNotifier: BspClientTestNotifier
+    private val bspClientTestNotifier: BspClientTestNotifier,
+    private val hasAnyProblems: Map<String, Set<TextDocumentIdentifier>>
 ) {
     private fun <T> withBepServer(body : (Server) -> T) :T {
-        val server = BepServer.newBepServer(compilationManager.client, compilationManager.workspaceRoot)
+        val server = BepServer.newBepServer(compilationManager.client, compilationManager.workspaceRoot, hasAnyProblems)
         val nettyServer = BepServer.nettyServerBuilder().addService(server).build()
         nettyServer.start()
         try {